+++ src/rules/no-empty-definitions.js
@@ -8,9 +8,28 @@
 //-----------------------------------------------------------------------------
 
 /**
- * @typedef {import("../types.ts").MarkdownRuleDefinition<{ RuleOptions: []; }>}
- * NoEmptyDefinitionsRuleDefinition
+ * @import { MarkdownRuleDefinition } from "../types.js";
+ * @typedef {"emptyDefinition" | "emptyFootnoteDefinition"} NoEmptyDefinitionsMessageIds
+ * @typedef {[{ allowDefinitions?: string[], allowFootnoteDefinitions?: string[], checkFootnoteDefinitions?: boolean }]} NoEmptyDefinitionsOptions
+ * @typedef {MarkdownRuleDefinition<{ RuleOptions: NoEmptyDefinitionsOptions, MessageIds: NoEmptyDefinitionsMessageIds }>} NoEmptyDefinitionsRuleDefinition
  */
+
+//-----------------------------------------------------------------------------
+// Helpers
+//-----------------------------------------------------------------------------
+
+const htmlCommentPattern = /<!--[\s\S]*?-->/gu;
+
+/**
+ * Checks if a string contains only HTML comments.
+ * @param {string} value The input string to check.
+ * @returns {boolean} True if the string contains only HTML comments, false otherwise.
+ */
+function isOnlyComments(value) {
+	const withoutComments = value.replace(htmlCommentPattern, "");
+
+	return withoutComments.trim().length === 0;
+}
 
 //-----------------------------------------------------------------------------
 // Rule Definition
@@ -29,16 +48,79 @@ export default {
 
 		messages: {
 			emptyDefinition: "Unexpected empty definition found.",
+			emptyFootnoteDefinition:
+				"Unexpected empty footnote definition found.",
 		},
+
+		schema: [
+			{
+				type: "object",
+				properties: {
+					allowDefinitions: {
+						type: "array",
+						items: {
+							type: "string",
+						},
+						uniqueItems: true,
+					},
+					allowFootnoteDefinitions: {
+						type: "array",
+						items: {
+							type: "string",
+						},
+						uniqueItems: true,
+					},
+					checkFootnoteDefinitions: {
+						type: "boolean",
+					},
+				},
+				additionalProperties: false,
+			},
+		],
+
+		defaultOptions: [
+			{
+				allowDefinitions: ["//"],
+				allowFootnoteDefinitions: [],
+				checkFootnoteDefinitions: true,
+			},
+		],
 	},
 
 	create(context) {
+		const allowDefinitions = new Set(context.options[0].allowDefinitions);
+		const allowFootnoteDefinitions = new Set(
+			context.options[0].allowFootnoteDefinitions,
+		);
+		const [{ checkFootnoteDefinitions }] = context.options;
+
 		return {
 			definition(node) {
-				if (!node.url || node.url === "#") {
+				if (
+					(!node.url || node.url === "#") &&
+					!allowDefinitions.has(node.identifier)
+				) {
 					context.report({
 						loc: node.position,
 						messageId: "emptyDefinition",
+					});
+				}
+			},
+
+			footnoteDefinition(node) {
+				if (
+					checkFootnoteDefinitions &&
+					!allowFootnoteDefinitions.has(node.identifier) &&
+					(node.children.length === 0 ||
+						node.children.every(
+							child =>
+								child.type === "html" &&
+								isOnlyComments(child.value),
+						))
+				) {
+					context.report({
+						loc: node.position,
+						messageId: "emptyFootnoteDefinition",
 					});
 				}
 			},