+++ tests/rules/no-empty-definitions.test.js
@@ -20,7 +20,7 @@ const ruleTester = new RuleTester({
 	plugins: {
 		markdown,
 	},
-	language: "markdown/commonmark",
+	language: "markdown/gfm",
 });
 
 ruleTester.run("no-empty-definitions", rule, {
@@ -29,6 +29,51 @@ ruleTester.run("no-empty-definitions", rule, {
 		"[foo]: #bar",
 		"[foo]: http://bar.com",
 		"[foo]: <https://bar.com>",
+		"[//]: # (This is a comment 1)",
+		"[//]: <> (This is a comment 2)",
+		"[^note]: This is a footnote.",
+		"[^note]: ![]()",
+		"[^note]: [text](url)",
+		"[^note]:\n    Content",
+		"[^note]:\n    > blockquote",
+		"[^note]: <span></span>",
+		"\\[^note]:",
+		"[\\^note]:",
+		"[^note\\]:",
+		"[^note]\\:",
+		"[^foo]: <span></span> <!-- comment -->",
+		"[^foo]: content <!-- comment -->",
+		"[^foo]: <!-- comment --> content",
+		"[^foo]: <!-- comment --> content <!-- comment -->",
+		dedent`
+		[^foo]: <!-- comm
+		    ent --> content <!-- comment -->
+		`,
+		{
+			code: "[foo]: #",
+			options: [{ allowDefinitions: ["foo"] }],
+		},
+		{
+			code: "[bar]: <>",
+			options: [{ allowDefinitions: ["bar"] }],
+		},
+		{
+			code: "[foo]: #\n[bar]: <>",
+			options: [{ allowDefinitions: ["foo", "bar"] }],
+		},
+		{
+			code: "[^note]:",
+			options: [{ checkFootnoteDefinitions: false }],
+		},
+		{
+			code: "[^note]:",
+			options: [
+				{
+					checkFootnoteDefinitions: true,
+					allowFootnoteDefinitions: ["note"],
+				},
+			],
+		},
 	],
 	invalid: [
 		{
@@ -74,6 +119,181 @@ ruleTester.run("no-empty-definitions", rule, {
 					column: 1,
 					endLine: 2,
 					endColumn: 10,
+				},
+			],
+		},
+		{
+			code: "[^note]:",
+			errors: [
+				{
+					messageId: "emptyFootnoteDefinition",
+					line: 1,
+					column: 1,
+					endLine: 1,
+					endColumn: 9,
+				},
+			],
+		},
+		{
+			code: "[^note]:   ",
+			errors: [
+				{
+					messageId: "emptyFootnoteDefinition",
+					line: 1,
+					column: 1,
+					endLine: 1,
+					endColumn: 12,
+				},
+			],
+		},
+		{
+			code: "[^note]:\n",
+			errors: [
+				{
+					messageId: "emptyFootnoteDefinition",
+					line: 1,
+					column: 1,
+					endLine: 1,
+					endColumn: 9,
+				},
+			],
+		},
+		{
+			code: "[^a]:\n[^b]:",
+			errors: [
+				{
+					messageId: "emptyFootnoteDefinition",
+					line: 1,
+					column: 1,
+					endLine: 1,
+					endColumn: 6,
+				},
+				{
+					messageId: "emptyFootnoteDefinition",
+					line: 2,
+					column: 1,
+					endLine: 2,
+					endColumn: 6,
+				},
+			],
+		},
+		{
+			code: "[foo]: #\n[^note]:",
+			errors: [
+				{
+					messageId: "emptyDefinition",
+					line: 1,
+					column: 1,
+					endLine: 1,
+					endColumn: 9,
+				},
+				{
+					messageId: "emptyFootnoteDefinition",
+					line: 2,
+					column: 1,
+					endLine: 2,
+					endColumn: 9,
+				},
+			],
+		},
+		{
+			code: "[foo]: #\n[^note]:",
+			options: [{ checkFootnoteDefinitions: false }],
+			errors: [
+				{
+					messageId: "emptyDefinition",
+					line: 1,
+					column: 1,
+					endLine: 1,
+					endColumn: 9,
+				},
+			],
+		},
+		{
+			code: "[^foo]: <!-- comment -->",
+			errors: [
+				{
+					messageId: "emptyFootnoteDefinition",
+					line: 1,
+					column: 1,
+					endLine: 1,
+					endColumn: 25,
+				},
+			],
+		},
+		{
+			code: dedent`
+			[^foo]: <!-- comment
+			    -->`,
+			errors: [
+				{
+					messageId: "emptyFootnoteDefinition",
+					line: 1,
+					column: 1,
+					endLine: 2,
+					endColumn: 8,
+				},
+			],
+		},
+		{
+			code: dedent`
+			[^foo]: <!-- comment -->
+			    <!-- another comment -->`,
+			errors: [
+				{
+					messageId: "emptyFootnoteDefinition",
+					line: 1,
+					column: 1,
+					endLine: 2,
+					endColumn: 29,
+				},
+			],
+		},
+		{
+			code: "[//]: #\n[foo]: #",
+			errors: [
+				{
+					messageId: "emptyDefinition",
+					line: 2,
+					column: 1,
+					endLine: 2,
+					endColumn: 9,
+				},
+			],
+		},
+		{
+			code: "[foo]: #",
+			options: [
+				{
+					allowDefinitions: ["bar"],
+					allowFootnoteDefinitions: ["foo"],
+				},
+			],
+			errors: [
+				{
+					messageId: "emptyDefinition",
+					line: 1,
+					column: 1,
+					endLine: 1,
+					endColumn: 9,
+				},
+			],
+		},
+		{
+			code: "[^foo]:",
+			options: [
+				{
+					allowDefinitions: ["foo"],
+					allowFootnoteDefinitions: ["bar"],
+				},
+			],
+			errors: [
+				{
+					messageId: "emptyFootnoteDefinition",
+					line: 1,
+					column: 1,
+					endLine: 1,
+					endColumn: 8,
 				},
 			],
 		},